Fix visualViewport page values in iframe
The visualViewport in an iframe should be entirely independent of the
visualViewport of the root frame. Unfortunately pageLeft/pageTop
erroneously add the root frame's viewport offsets to the document
offset, even when inside an iframe. This CL fixes the issue so that
iframe's report only the document offset (since an iframe cannot be
zoomed, its visualViewport cannot have its own offset).

Bug: 1313970
Change-Id: I917bf72f645d105cdde86bf186e7bebd3741247e