Handle the context getting discarded in NormalizeAlgorithm
The WebCrypto result machinery tracks when the context was discarded
to cancel the operation, if still pending. However, that logic did not
correctly account for the context being discarded before the task is
posted:

- CryptoResultImpl::Resolver DCHECKs that the context is valid when
  it is created. As far as I can tell, that DCHECK isn't necessary?

- If CryptoResultImpl::Cancel is called before calling
  CryptoResultImpl::Result, the cancel flag is null, which breaks
  WebCryptoResult's DCHECKs and null pointer derefs later in release
  code.

Instead, remove the DCHECK and retain the cancel flag in CryptoResultImpl
so WebCryptoResult can observe cancelation in either order.

These two changes fix the crashes. This CL additionally makes
webcrypto_impl.cc check for cancellation before posting the task. This
isn't strictly necessary, but avoids a pointless thread hop.

Bug: 1324846
Change-Id: Idd8f2174f438fc3bfb5ffd140b9983d5b6b8039d