Replace setTimeout with rAF in the select-event.html test
The previous code used a 200ms timeout to let events fire, but that made the test take a very long time (~34s for all browsers). Two rAF() loops are now used instead, which helped reduce the total test time to ~3.4s.

Another change in this CL is to use `element.addEventListener` to uniformize all unit tests within this test file.

The remaining failed tests are not due to timeout errors and hence, not fixed as part of this CL.

Bug: 1161590

Change-Id: I150dfbe0b381599f3d24da94a2e2f4af6a6b6f75